Converted networkInfo to hooks
